requirements.txt
customtkinter>=5.2.0
PyMuPDF>=1.23.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
scipy>=1.7.0
pyinstaller>=5.0.0
